Add changes for first incremental rebuild
In an effort to track down a segfault, I backed out all changes I've
added in the last few months.  Basically, I started with the version
that works with the calculation of vertical shear that I added, and
I'm going to build back, piece by piece.  For this commit, I have
added in just the changes necessary to add new arrays for the new
variables & diagnostics that I'll be adding.  Note that this includes
allocate statements for these arrays, which is critical for this
debugging effort, since it's a segfault.  I added the changes
necessary for reading in data in getdata_grib and getdata_netcdf.

I tested this version and it works fine to completion with no
segfault.  Yippee!

4/15/2022