Add changes for third incremental rebuild
In this increment, I added in just the beginning of the code for
subroutine get_gen_diags, including subroutines get_divg, divcal,
get_smooth_value_at_pt, and sst_barnes.

I tested the version and it works find to completion with no segfault.
Additionally, the results match with the previous increment.

4/19/2022