Add changes for SUCCESSFUL fix
In this incremental change, I fixed the issue with the segfault in
subroutine output_atcfunix.  The problem was that in the call to this
subroutine, I would pass in -99 or -999 values for parameters that
were undefined (e.g., in the case of the tracker being unable to find
a storm at 00h).  I have actually done this for many years, but the
problem currently was that I was trying to then modify that value that
had been sent to the subroutine.  This is a Fortran no-no!  When sending
a numerical value as an argument to a subroutine, Fortran interprets
that as a constant, and then when I would go to modify it in the
subroutine, I'm treating it like a variable -- BANG!... segfault!  So
what I did is that for all calls to output subroutines, like
output_atcfunix, output_aext, output_atcf_gen, etc., I created variables
for holding -99 or -999 values and I make sure to pass those to the
subroutines instead.  This is the only issue that was causing the
segfault.  Done.  Good riddance!

4/26/2022