Fix R34 code and fix distance bug and separate SST
For the new R34 code, I was getting some wonky results because I was interpolating between the nearest two radial bins as I did with the previous R34 code.  The problem is that in the old R34 code the distances were sorted and so the interpolation was possible, but in this new R34 code that interpolation fails because the radial bins are not sorted.  So we remove the interpolation and just use the value at the radial bin, which is fine because the radial bin resolution is 3 km which is fine enough already.  Also in this version I fixed a bug in subroutine calcdist, which is the main distance calculation routine in the tracker.  Fortunately, this bug would have only appeared in unusual cases where distance was being calculated between two extremely close points.  I fixed this in calcdist by using double precision variables and double precision versions of trig functions.  In this version, I also separated out the computation of area-averaged SST from the genesis stuff.  This was done so that the SST could be computed and written out for all cases, not just genesis cases.