Fix bug in allocating SST array
I fixed a bug in gettrk_main.f (brought to my attention by Matt Morin).
I found that  the sst array was being used without first being allocated.
This was because I had the allocate statement within an IF statement
controlled by the value of genflag, where instead I needed a separate
IF statement controlled by the value of sstflag (this comes from the
namelist).  I also fixed a couple of basin IDs for the output basin IDs.