Add changes to relax R34 checking restrictions
Testing suggested that the new R34 scheme may be too restrictive / conservative in its QC checking, whereby some cases were showing outer rainbands with healthy winds exceeding 34 kts getting screened out.

Here is what I changed in the source.  After an R34 has been diagnosed, there are checks that must be passed in order for that R34 to be reported and for the R34 search to successfully stop.  This is done via ensuring that the mean Vt at the diagnosed R34 radius is at least 0.5 * Holland_profile at that radius based on the tracker-derived Vmax and tracker-derived axisymmetric RMW.  This check must be passed for 60% of the bins (previous version: 75% of the bins) over a range of distances moving radially inward from the diagnosed R34.  In the previous version I gave you, that distance was a percentage, 10%, of the diagnosed R34 (e.g., if R34 = 250 km, then that check must pass for 75% of the bins moving radially inward over 25 km).  In the current version, I have it as a fixed distance that is input (for testing purposes) via namelist.  In that namelist I gave you, you'll see a new parameter called radii_width_thresh, which in my testing I have set to 15.0 (units = km).  So that 15 km spans exactly 5 of the 3-km radial bins, and as I just mentioned, that check must pass for 60% of the bins moving radially inward... or 3 out of 5.

Additionally, I added code that will allow for a "free pass" for the R34 if the Vt signature near the diagnosed R34 is robust enough.  With this "free pass", no further checking is done with all of that Holland profile stuff, and the diagnosed R34 is taken as found.  For the first option, if the mean Vt of the radial band in that quadrant exceeds 34 kts, then it's a no-brainer -- give the free pass and report the R34 as found without further checking.  For a second option, a percentile wind is checked for 34-kt wind exceedance, and if that passes, then a free pass is issued.  As an example, in the input namelist I included, I set another new parameter, radii_free_pass_pctile = 67.0.  So if the 67th percentile wind is at least 34 kts (or said another way, if at least 1/3 of the winds in this radial band have a wind value > 34 kts), then a free pass is issued and no further checking is done.  So there are two additional "knobs" that can be adjusted here via the namelist to help tune the performance while we are testing.  If neither of these free passes holds true, then the further checking described above with the Holland profile is performed to confirm that the diagnosed R34 can be considered part of the circulation of the storm.