Add changes to fix Greenwich Meridian crossing
The main change of this version is to add a lot of code to handle the
problem of the HAFS grid straddling across the Greenwich Meridian.
Code was added to bump up the longitudes so that the grid lon "min"
was no longer greater than the grid lon "max" (e.g., the HAFS grid
was going from a grid lon min of 355E to a grid lon max of 100E, and
the problem was compounded by the fact that TC Vitals in the South
Indian Ocean are in 0-180E space.  Another change here was to add in
Biju's change to the namelist, where instead of it being piped in
from the script, it is being opened in subroutine read_nlists as
unit 555.