Fix additional GM-wrapping issues
The main component of this update is a an additional fix for
grids that span across the Greenwich Meridian (GM).  Specifically,
there was an issue with output from subroutine distbear, where
the target longitude was not in the same reference space as that
of the point of origin (the point of origin had longitudes that
were extending beyond 360, while targlon is always returned from
subroutine distbear in 0-360 space, not 360+ space).  An additional
change was made to subroutine output_atcfunix to correct ocean
basin IDs that had been messed up.