Make light changes to sync with Biju's HAFS version
The one significant change was to set the variable
shear_dir_point_to = -999.0 in subroutine get_shear.  This is
in case the wind levels for shear are not available, and then
this variable will be initialized to a missing value of -999.0,
and this initialization avoids problems later of using an
uninitialized variable.  Other than that, a couple of comments
were removed in order to exactly match Biju's version that was
submitted with the HAFS package.