Fix sleeper-cell verbose setting bug
I call this a sleeper-cell bug because if the user keeps the verbose
setting in the namelist as 3, then this bug will never appear.  But if
a user sets the verbose flag to something less than 3, then certain bugs
will appear, because meaningful executable code was erroneously included
inside of the IF statements for the verb>=3 checks.  This occurred in a
number of locations in the code.  In this update, I have fixed all of
these instances.