Add L-S mask capability for GRIB1 plus ECMWF gen
There are three key updates in this commit: (1) The ability to use
a separate land-sea mask file for GRIB1 files was added.  This functionality
was already included for NetCDF files.  This is useful in the event that a
user is tracking with data for a model that doesn't include a land-sea mask
in their files; (2) Fixed the GRIB1 PDS IDs for ECMWF data that are needed
for the extra TC genesis diagnostics (these IDs are different from the IDs
that NCEP uses for its variables, i.e., a different Parameter Table 2 is
used); and (3) A bug was fixed related to the use of uninitialized
variables.  There were certain scalar integer variables that I was using
to check return codes from deallocate statements.  The deallocate
statements were never getting executed, so the variables were never getting
set, but then I was checking the values of those integers for errors, and
with the Intel compiler on WCOSS2 and the compiler options I was using,
those uninitialized variables had wacky values that were triggering a
STOP statement in the IF checks.  I fixed this to ensure that I'm
initializing these variables before using them.