Update old version of tave with numerous changes
The version of tave.f in this repo was a much older version that
was missing key changes to address critical bugs.  These include:
(1) Changing the variable ndpts to be ngrdpts, as this was giving
an incorrect number of returned data points; (2) The gf_free
function from the NCEP g2lib was missing.  Without this, memory
does not get freed up after grib2 reads and the executable starts
racking up additional memory during; (3) A fix was made to how
a temporary array is named and used, which makes a difference in
how the data are handled, and (4) A critical error was fixed that
was preventing temperature data from being read in and also the
lead time from being read in, in both cases leading to incorrect
data being read in.