Update old version of vint with numerous changes
The version of vint.f in this repo was a much older version that
was missing key changes to address critical bugs.  These include:
(1) Changing the variable ndpts to be ngrdpts, as this was giving
an incorrect number of returned data points; (2) The gf_free
function from the NCEP g2lib was missing.  Without this, memory
does not get freed up after grib2 reads and the executable starts
racking up additional memory during execution, and (3) A fix was
made to how a temporary array is named and used, which makes a
difference in how the data are handled.