Add adjustment for MSLP contour interval units
An issue was found with the units used for the MSLP contour interval
that a user requests in the namelist, specifically for the value that
gets used for trkrinfo%contint.  The value of those units must match
between the units in the input data file and what the user requests.
In other words, if the data file has MSLP stored in mb, then the
user request for trkrinfo%contint must also be in mb.  I added code
in subroutine check_closed_contour to ensure this happens and make
any unit adjustments for trkrinfo%contint from mb <---> Pa if
necessary.