Change units of istmspd to knots from knots x ten
In the ouptut atcfunix record, the tracker-diagnosed translation speed of
the model storm motion is output.  Since the day I first switched from the
old ATCF format to the new atcfunix format in 1999, I've been using a value
for the units of translation speed as knots*10.  This was following the
guidance from NHC on the format at the time.  At some point since then, the
format has changed so that the units are now just knots.  I had to modify
the code in two locations to make this compliant.  The value is stored in
the istmspd variable before being written out.