Apply step 1 of grid edge workaround
An issue was raised by Sam at GSL with MPAS which turned out to be very similar
to an issue with tracking with HWRF back in 2020.  The issue with both HWRF and
MPAS is that their post-processing systems are inconsistent in how they handle
the grid edges.  For some variables, the edge points with valid and missing
data differ from those in other variables.  For MPAS, this was showing up for
10-m winds and MSLP being different from U850 and V850.  I added a second level
of checking in addition to the logical bitmap in order to fix this.  In this
second level, I also check for the initialized value of -9999.