Add check for invalid data to bilinear interp
As part of the fix to address inconsistent grid edges in model data
among momentum and mass fields due to the interpolation from native
grid to output grid that occurs in their post-processors, I added a
check for the -9999 initialization value that works in tandem with
the check for the bitmap flag.  I added this specific check to the
bilin_int_uneven subroutine.