Fix vortex-tilt errors found during SHiELD testing
A lot of new code was added to diagnose vortex tilt using either
relative vorticity, wind circulation, gp height or temperature.
These changes were tested using both SHiELD and T-SHiELD NetCDF
data, and a number of fixes had to be made to the code.  In
addition, a change had to be made to the cwaitfor.c program to
fix an output formatting issue that only appeared once we
compiled with gnu instead of intel.