Add output of u-component of shear
In response to a request, I added the output of the u-component of the
850-200 mb vertical shear.  This was easy to do because the vertical
shear is already computed in the tracker (including the separate u- and
v-components), and so this just required the output of that u-component.
Currently, this is only output in the atcf_gen output file that is used
for genesis detection & tracking.