Fix allocatable arrays and allow EMC unit numbers
This change fixes a bug that was found by Alan Brammer where allocatable
arrays that were not yet allocated were being accessed, leading to a
segfault.  This was happening for the sst array if sstflag was not set
to Y, and also for the q850, rh, spfh, temperature and omega500 arrays
if genflag was not set to Y.  This has been fixed by adding relevant IF
statements.  Also, in order to accommodate the EMC method of specifying
unit numbers (with environmentals variables like "FORT51", as opposed to
something like "fort.51"), code was added by Hana Jafary from EMC for
the open statements in tave, vint and gettrk.  So now it can handle open
statements using either method.